% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustics_datasets_directory.R
\name{Acoustics_datasets_directory}
\alias{Acoustics_datasets_directory}
\title{Set the directory of the acoustic data used by the sonR package}
\usage{
Acoustics_datasets_directory(path = NULL)
}
\arguments{
\item{path}{The path to the directory in which to put acoustic data, structured as CruiseName/"Events"/EventName/EchosounderName/"raw" for the raw files and substituting "raw" by "tsd" for TSD files.}
}
\description{
Set the directory of the acoustic data used by the sonR package
}
