% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.school.R
\name{cm.school}
\alias{cm.school}
\title{Calculates the center of mass of a (simulated) school in 1D, 2D or 3D, or based on acoustic data.}
\usage{
cm.school(school, mass = 1, subset = NULL, na.rm = TRUE,
  excl.neg = TRUE)
}
\arguments{
\item{school}{is either a list of x-, y- and z-positions and mass of the fish (named "psxf", "psyf", "pszf" and "mass"), a list of x-, y- and z-positions and volume and acoustic intensity of the voxels (named "psxx", "psyx", "pszx", "volx" and "vbsc" or "mvbs"), or a matrix of columns holding the positions along each dimension.}

\item{mass}{is the mass of the fish.}

\item{subset}{is a numeric or logical vector defineing which positions to include in the calculation of center of mass.}

\item{na.rm}{is used in sum().}

\item{excl.neg}{is FALSE if negarive mass values are to be included in the calculation of the centers of mass of the school.}
}
\value{

}
\description{
Calculates the center of mass of a (simulated) school in 1D, 2D or 3D, or based on acoustic data.
}
\examples{
\dontrun{}

}
