% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.events.R
\name{combine.events}
\alias{combine.events}
\title{(Internal) Merge segmentation events, used in SX90.segment.event() and extract_event().}
\usage{
combine.events(
  events,
  event,
  cruise = NULL,
  esnm = NULL,
  dir.data = NULL,
  ctd = 1
)
}
\arguments{
\item{events}{is a list of beam configuration data.}

\item{event}{is a list of beam configuration data.}

\item{cruise}{is a list of beam configuration data.}

\item{esnm}{is a list of beam configuration data.}

\item{dir.data}{is a list of beam configuration data.}

\item{ctd}{is a list of beam configuration data.}
}
\description{
(Internal) Merge segmentation events, used in SX90.segment.event() and extract_event().
}
\examples{
\dontrun{}

}
