% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compr.event.R
\name{compr.event}
\alias{compr.event}
\alias{compr.event_oneFile_write}
\alias{compr.TSD}
\title{This funciton compresses the data stored in 'x' by averaging (either mean, trimmed mean or median) in depth and time bins. This is designed primarily for echosounder beams.}
\usage{
compr.event(
  event,
  filenr = "all",
  tres = NULL,
  xres = NULL,
  zres = NULL,
  rres = NULL,
  bres = NULL,
  cores = 1,
  funvbsc = c("mean", "median"),
  funt = c("median", "mean"),
  adds = NULL,
  split = TRUE,
  skipAngles = TRUE,
  origin = 1,
  z0 = 0,
  cruise = NULL,
  esnm = "EK60",
  msg = TRUE,
  write = TRUE,
  filesize = 3e+08,
  chunksize = 3e+08,
  clear_individual = TRUE,
  clear_along = FALSE,
  maxlenb = NULL,
  ow = FALSE,
  ...
)

compr.event_oneFile_write(
  i,
  indt,
  filelist,
  pingsfiles,
  vesselfiles,
  beamsfiles,
  t = "all",
  drop = TRUE,
  compress = FALSE,
  TIME = NULL,
  write = TRUE,
  tres = NULL,
  xres = NULL,
  zres = NULL,
  rres = NULL,
  bres = NULL,
  funvbsc = c("mean", "median"),
  funt = c("median", "mean"),
  adds = list(),
  split = TRUE,
  skipAngles = TRUE,
  origin = 1,
  z0 = 0,
  pingsnames = NULL,
  vesselnames = NULL,
  beamsnames = NULL,
  dumpfiles = NULL,
  maxlenb = NULL,
  ...
)

compr.TSD(
  data = NULL,
  tres = NULL,
  xres = NULL,
  zres = NULL,
  rres = NULL,
  bres = NULL,
  funvbsc = c("mean", "median"),
  funt = c("median", "mean"),
  adds = list(),
  split = TRUE,
  skipAngles = TRUE,
  origin = 1,
  z0 = 0,
  drop = FALSE,
  ...
)
}
\arguments{
\item{tres}{The time resolution of the compressed data in seconds.}

\item{xres}{The sailed distance resolution of the compressed data in meters.}

\item{zres}{The depth resolution of the compressed data in meters.}

\item{rres}{The range resolution of the compressed data in meters.}

\item{bres}{The beam resolution of the compressed data in integer number.}

\item{cores}{is an integer specifying the number of cores to run the compression over in parallel (should be ler than the number of cores in the computer).}

\item{funvbsc}{is the function to apply in the compression, either given as function or as a string, in which case the strings "mean" and "median" represents fast versions of the functions with the corresponding names (sum()/length() and fastMedian(), respectively). Default is mean, which is recommended for volume backscattering coefficient (sv) data, which are by nature exponentially distributed (backscattering from multiple targets), and using median will underestmate the true backscatter.}

\item{funt}{is the same as funvbsc, but used for averaging vessel data in the new time/distance bins.}

\item{adds}{is a list of additional data overriding corresponding variables in 'data'}

\item{split}{used in psx.TSD().}

\item{skipAngles}{is TRUE to discard electircal angles from the data (saves time).}

\item{origin}{is either the time index of the origin, or the origin itself, given as c(longitude, latitude).}

\item{write}{is FALSE to only return the data and not write to TSD file.}

\item{ow}{Logical: If TRUE, overwrite the latest compression (directory).}

\item{...}{further arguments passed to psx.TSD().}

\item{data}{is the list of inputs variables as returned from EKRaw2TSD().}
}
\value{



}
\description{
This funciton compresses the data stored in 'x' by averaging (either mean, trimmed mean or median) in depth and time bins. This is designed primarily for echosounder beams.

(Internal) Read and write one Simrad raw file.

This funciton compresses the data stored in 'x' by averaging (either mean, trimmed mean or median) in depth and time bins. This is designed primarily for echosounder beams.
}
\examples{
\dontrun{}

\dontrun{}

}
\keyword{internal}
