% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\title{Returning a categorical array (0 and 1) for the condition defined by 'a', 'b' and 'cond'.}
\usage{
contains(x, a, b = NULL, cond = "less")
}
\arguments{
\item{x}{is the input array.}

\item{a}{is the lower border given as a single value or as a vector c(a,b).}

\item{b}{is the upper border. If 'b' is a character, it is interpreted as 'cond'..}

\item{cond}{specifies the border condition type:}
}
\value{

}
\description{
Returning a categorical array (0 and 1) for the condition defined by 'a', 'b' and 'cond'.
}
\examples{
\dontrun{}

}
