% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctVessel.R
\name{correctVessel}
\alias{correctVessel}
\title{Shifts the vessel information 'rtzv', 'lonv', 'latv', and 'ispv' by the time offset 'TOV'. Used in read.event().}
\usage{
correctVessel(var, time, rawvessel, TOV = 0)
}
\arguments{
\item{var}{is a vector of the names of the list holding the vessel data (obtained by names(out) in read.event()).}

\item{time}{is a list of the time information (obtained by out[c("mtim","utim")] in read.event()).}

\item{rawvessel}{is a vector of the paths to the TSD files from which the data are read.}

\item{TOV}{is the time offset of the vessel information, caused by some error in the SIMRAD system.}
}
\value{

}
\description{
Shifts the vessel information 'rtzv', 'lonv', 'latv', and 'ispv' by the time offset 'TOV'. Used in read.event().
}
\examples{
\dontrun{}

}
