% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTD2TSD.R
\name{CTD2TSD}
\alias{CTD2TSD}
\title{Converts a CTD file to the TSD format for the variables relevant to echoIBM().}
\usage{
CTD2TSD(f = NULL, outfile = NULL, hpr0 = NULL, Pain = FALSE, vessel = list())
}
\arguments{
\item{f}{is the CTD file in cnv format.}

\item{outfile}{is file to which the CTD data are written.}

\item{hpr0}{is the pressure at sea level, which must be given externally. Defaulted to 101325 pascal.}

\item{Pain}{is TRUE if pressure "ihpr" is given in Pascal and FALSE if given in decibar relative to surface pressure (10000 Pascal, giving values approximately equivalent to water depth).}

\item{vessel}{A list of vessel info by which the closest CTD station is found as the minimum of the sum of distance in meters and seconds.}
}
\value{

}
\description{
Converts a CTD file to the TSD format for the variables relevant to echoIBM().
}
\examples{
\dontrun{}

}
