% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTDtimeandpos.R
\name{CTDtimeandpos}
\alias{CTDtimeandpos}
\title{Reads the time and position of a CTD file in the cnv format.}
\usage{
CTDtimeandpos(x)
}
\arguments{
\item{x}{is a CTD file in the cnv format.}
}
\value{

}
\description{
Reads the time and position of a CTD file in the cnv format.
}
\examples{
\dontrun{}

}
