% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.pdns.R
\name{echoIBM.pdns}
\alias{echoIBM.pdns}
\title{Calculates the periodic noise for all voxels based on estimates of the parameters of the periodic noise for ONE SINGLE TIME STEP. The function assumes that the phase is correctly given in data$pns3.}
\usage{
echoIBM.pdns(data, indt = NULL, TVG = FALSE, TVG.exp = 2)
}
\arguments{
\item{data}{is a list containing beam configuration, and periodic noise parameters. Speficically the following variables must be included: Beams variables: 'sint', 'lenb'; periodic noise variables: 'pns1', 'pns2', 'pns3', 'acfq', 'harm'; and either 'bgns' or 'numb'.}

\item{indt}{is the time step identifier, which if given selects the phase parameter from the matrix data$pn3M.}

\item{TVG}{is TRUE if TVG should be added to the periodic noise.}

\item{TVG.exp}{is the exponent of the eamotric spreading of the sound wave, theoretically 2 for Sv and 4 for TS.}
}
\value{

}
\description{
Calculates the periodic noise for all voxels based on estimates of the parameters of the periodic noise for ONE SINGLE TIME STEP. The function assumes that the phase is correctly given in data$pns3.
}
\examples{
\dontrun{}

}
