% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bseg.event.R, R/pseg.event.R
\name{bseg.event}
\alias{bseg.event}
\alias{pseg.event}
\title{NA}
\usage{
bseg.event(
  event = 1,
  t = 1,
  cruise = 2009116,
  bgns = NULL,
  beta0 = NULL,
  beta1 = NULL,
  misM = NULL,
  iter = TRUE,
  factor = 10^(-5/10),
  minmsvM = 10^(-40/10),
  pow = 1,
  ind = list(-(1:30), NULL),
  nsind = 0.75,
  smind = list(-(1:300)),
  range = list(),
  subset = NULL,
  h = NULL,
  alpha = NULL,
  filesize = 3e+08,
  hins_add = 10,
  turns = 10,
  phase = TRUE,
  TVG.exp = 2,
  dens = NULL,
  segdir = NULL,
  esnm = "MS70",
  subtractNoise = TRUE,
  pr0s.out = FALSE,
  adds = list(),
  sim = TRUE,
  allow.old = FALSE,
  TOV = 0,
  startn = 1,
  prefix = "",
  ...
)

pseg.event(
  event = 1,
  t = 1,
  cruise = 2009116,
  alpha = NULL,
  filesize = 3e+08,
  dens = NULL,
  segdir = NULL,
  adds = list(),
  TOV = 0,
  startn = 1,
  ...
)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{t}{is either the number of the ping to be treated, as listed from 1 to the number of pings in the event, or the time point given as a string "yyyymmddHHMMSS.FFF" or "HHMMSS.FFF". Only one time step alowed!}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event.}

\item{bgns}{is ann optional list of noise estimates (must contain the background noise 'bgns', and may contain the periodic noise estimates 'pns1', 'pns2' and 'harm'). The phase parameter 'pns2' is extracted from the data.}

\item{beta0}{is the minimum schooling threshold for the volume backscattering coefficient. For voxels where beta0<noise: beta0=noise.}

\item{beta1}{is the maximum schooling threshold for the volume backscattering coefficient, defining the probability distribution of the signal. Should be chosen on the basis of the maximum packing density of the observed species.}

\item{misM}{is the mean expected observed volume backscattering coefficient (linear value) used in the segmentation. Else, 'misM' should have length equal to the number of time steps. If misM==NULL the absolute lower and upper schooling threshold 'lsth' and 'usth' are used.}

\item{iter}{is TRUE to apply the iterative method where 'misM' is used in an initial segmentation, the result of which is used to re-estimate 'misM' by the funciton meanSv.TSD() which is then used in the final segmentation. If more than one time steps is requested, the previous estimated 'misM' is used as the initial estimate for the next time steps (thus the input value of 'misM' disregarded). If 'turns' is given different from 1, the mean of the currently processed block of pings is used as the initial estimate of 'misM' for all pings of the next block of pings.  The iterative method uses 'misM' as an initial estimate used when segmenting for 5 dB below this value, and then the 'misM' is estimated from the resulting segmentation mask (scaled by a factor determined from simulations), and then segmenting again with the new estimate.}

\item{factor}{is the factor by which the input mean observed volume backscattering strength is multiplied to obtain the enlarged segmentation mask. Lower value normally leads to larger segmentation mask.}

\item{minmsvM}{is only used when the length of 'misM' is 1 (iterative estimation of the mean sv), and ensures that 'misM' does not go below 'minmsvM' (preventing surface scattering and other noisy values).}

\item{pow}{is a value to raise the segmentation threshold values to the power of, which if pow<1 eases the strictness of the segmentations, suitable for bottom detection (pow=1/4 might be good):}

\item{ind}{is a vector or list of indices along the beams, as input to ind.expand(), used to select the subset over which the estimation of background noise is done. Also voxels not included in this subset are assigned p=1.}

\item{nsind}{is a vector of indices along the beams, as input to ind.expand(), used to select the subset over which the estimation of the phase of the periodic noise is done. If given as a single numeric, the outermost 'nsind' voxels are used in each beam.}

\item{smind}{is a list of indices used to define the subset which is smoothed spatially to accumulate the probabilities. The default avoids smoothing densely packed voxels, which reduces computational time.}

\item{h}{is the bandwidth of the spatial Gaussian kernel smoother, given as the standard deviation of the Gaussian distribution. For no smoothing use h=NULL.}

\item{alpha}{is the significance level of the hypothesis testing of H0: school not present in the voxel, against H1: school present in the voxel. To return non-thresholded data, use alpha=NULL.}

\item{hins_add}{is the number of voxels that should be discarded on both sides of high intensity noise voxels voxels along beams, used for accounting for possible high values that are related to the high intensity noise but not classified as such voxels.}

\item{turns}{is the number of time steps treated in each block of time steps, where higher values demand more memory but reduces CPU time.}

\item{phase}{is FALSE if any of 'pn3M' (phase for each time step) or 'pns3' (phase equal for all time steps) given in 'bgns' or read from the noise file located by the funciton noise.path.event() should be used, as oposed to estimating the phase from the data for each time step. This is only recommended for simulated data where the phase is constant over all time steps, and saves some CPU time.}

\item{dens}{is the packing density of the school, used when var="posf". These values are attemptedly read in the event, if not given as input to the function.}

\item{segdir}{is an optional string giving the name of the folder in which to put the segmentation files. The folder will be put in the directory of the event. segdir==NULL results in segdir="seg0".}

\item{subtractNoise}{is TRUE if the original acoustic data should be subtracted noise, which is used when returning the total and mean volume backscatter from the segment.}

\item{pr0s.out}{is TRUE if the smoothed probabilities of insuffuciently high true volume backscattering coefficient to imply school should be returned in arrays of the same dimensions as the volume backscattering coefficient 'vbsc'.}

\item{adds}{is an optional list of variables overriding the variables in read from the event.}

\item{sim}{is a TRUE if smoothing should be done only along the first dimensions, simultaneously over the stages of the last dimension. If 'sim' is an integer larger than 1, the positions 'coords' are used 'sim' times, and the data 'x' should have length 'sim' times the length of one coordinate of 'coords'.}

\item{dir.data}{is the path to the directory in which the projects are stored, defaulted by the variable Acoustics_datasets_directory().}

\item{var}{is "vbsc" for segmenting volume backscattering data, and "posf" for segmenting original fish position data.}
}
\value{



}
\description{
NA

Segment the original fish positions as the voxels that contain at least the fraction 'alpha' of the expected number of fish, calculated as the volume of the voxels times the packing density:
}
\examples{
\dontrun{}

\dontrun{}

}
