% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bseg.TSD.R
\name{echoIBM.vbsc2p.event}
\alias{echoIBM.vbsc2p.event}
\title{Estimates the probability of not school in voxels of sonar data.}
\usage{
echoIBM.vbsc2p.event(
  event = 1,
  t = 1,
  cruise = 2009116,
  bgns = NULL,
  beta0 = NULL,
  beta1 = NULL,
  factor = NULL,
  ind = list(-(1:30), NULL),
  nsind = 0.75,
  smind = list(-(1:300)),
  range = list(),
  subset = NULL,
  h = NULL,
  alpha = NULL,
  dir.data = NULL,
  hins_add = 10,
  phase = TRUE,
  TVG.exp = 2,
  esnm = "MS70",
  subtractNoise = TRUE,
  adds = list(),
  sim = TRUE,
  na.rm = 1,
  allow.old = FALSE,
  TOV = 0
)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{t}{is either the indexes of the pings to be treated, as listed from 1 to the number of pings in the event, or the time point given as a string "yyyymmddHHMMSS.FFF" or "HHMMSS.FFF".}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event.}

\item{bgns}{is an optional list of noise estimates (must contain the background noise 'bgns', and may contain the periodic noise estimates 'pns1', 'pns2' and 'harm'). The phase parameter 'pns2' is extracted from the data. Use 'bgns' to suppress adding periodic noise or near range noise in the noise which is compared to the data, by setting bgns=list(acfq=NULL) or bgns=list(nr0a=NULL).}

\item{beta0}{is the minimum schooling threshold for the volume backscattering coefficient. For voxels where beta0<noise: beta0=noise.}

\item{beta1}{is the maximum schooling threshold for the volume backscattering coefficient, defining the probability distribution of the signal. Should be chosen on the basis of the maximum packing density of the observed species.}

\item{factor}{is one or several factors to multiply the schooling thresholds 'beta0' and 'beta1' by, increasing the memory occupied by the function by 50 \% of the original memory (factor==NULL) for each element of 'factor'. This option was included to reduce CPU time in the case that both an unbiased and an enlarged segmentation mask is requested.}

\item{ind}{is a vector or list of indexes along the beams, as input to ind.expand(), used to select the subset over which the estimation of background noise is done. Also voxels not included in this subset are assigned p=1.}

\item{nsind}{is a vector of indexes along the beams, as input to ind.expand(), used to select the subset over which the estimation of the phase of the periodic noise is done. If given as a single numeric, the outermost 'nsind' voxels are used in each beam.}

\item{smind}{is a list of indexes used to define the subset which is smoothed spatially to accumulate the probabilities. The default avoids smoothing densely packed voxels, which reduces computational time.}

\item{h}{is the bandwidth of the spatial Gaussian kernel smoother, given as the standard deviation of the Gaussian distribution. For no smoothing use h=NULL.}

\item{alpha}{is the significance level of the hypothesis testing of H0: school not present in the voxel, against H1: school present in the voxel. To return non-thresholded data, use alpha=NULL.}

\item{dir.data}{is the path to the directory in which the projects are stored, defaulted by the variable Acoustics_datasets_directory().}

\item{hins_add}{is the number of voxels that should be discarded on both sides of high intensity noise voxels voxels along beams, used for accounting for possible high values that are related to the high intensity noise but not classified as such voxels.}

\item{phase}{is FALSE if any of 'pn3M' (phase for each time step) or 'pns3' (phase equal for all time steps) given in 'bgns' or read from the noise file located by the funciton noise.path.event() should be used, as oposed to estimating the phase from the data for each time step. This is only recommended for simulated data where the phase is constant over all time steps, and saves some CPU time.}

\item{TVG.exp}{is the exponent of the eamotric spreading of the sound wave, theoretically 2 for Sv and 4 for TS.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{subtractNoise}{is TRUE if the original acoustic data should be subtracted noise, which is used when returning the total and mean volume backscatter from the segment.}

\item{adds}{is an optional list of variables overriding the variables in read from the event.}

\item{sim}{is a TRUE if smoothing should be done only along the first dimensions, simultaneously over the stages of the last dimension. If 'sim' is an integer larger than 1, the positions 'coords' are used 'sim' times, and the data 'x' should have length 'sim' times the length of one coordinate of 'coords'.}

\item{na.rm}{is single integer representing the dimension along which NAs are discarded from the smoothing in the case that sim==1. For example, if na.rm=2 and the dimension of 'x' is [5,12,7], and x[3,2:4,5]=NA, then all data x[,2:4,] will be excluded from the smoothing and set to NA. If na.rm=FALSE, no NAs should be contained in the data.}
}
\description{
Estimates the probability of not school in voxels of sonar data.
}
\examples{
\dontrun{}

}
