% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgedir3D.R
\name{edgedir3D}
\alias{edgedir3D}
\title{Calculates the edge points of an acoustic device based on the CTD-data, using edge points along the beam maxima as basis, and extracting the vertical dimension (elevation angle) from the vertical angular distance between beam maxima. !!!Only one time step is treated!!!!}
\usage{
edgedir3D(
  data,
  esnm = NULL,
  seabed = -12000,
  rot = 1,
  compensation = c("pitch", "roll"),
  ideal = TRUE,
  ...
)
}
\arguments{
\item{data}{is the list of TSD inputs as returned from read.TSD (must contain ""dira" and "dire", and CTD-data).}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{seabed}{is the z-coordinate of the sea bed, usually provided by echo sounder data. Soundbeams reflected from the sea bed or the surface are reflected at the incidence angle.}

\item{rot}{is 1 if simple rotation using cosine and sine is to be used, and 2 if the function rotate() is to be used in the rotation. Times for the different methods (tested on MacBook Pro dual 2.8 GHz, 2010-02-09, with other applications running):}

\item{compensation}{is a vector of string giving which rotation values that are compensated for in the sonar. Only c("pitch","roll") is available for the current version. Used in soundbeam.TSD.}

\item{ideal}{is TRUE to represent the simple case where the speed of sound 'data$asps' is invariant of depth.}

\item{...}{Used for robutness.}
}
\description{
Calculates the edge points of an acoustic device based on the CTD-data, using edge points along the beam maxima as basis, and extracting the vertical dimension (elevation angle) from the vertical angular distance between beam maxima. !!!Only one time step is treated!!!!
}
\examples{
\dontrun{}

}
