% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EKRaw2TSD_getesnm.R
\name{EKRaw2TSD_getesnm}
\alias{EKRaw2TSD_getesnm}
\title{(Internal) Get the name of the acoustic system form the raw data.}
\usage{
EKRaw2TSD_getesnm(thisd, numt)
}
\value{

}
\description{
(Internal) Get the name of the acoustic system form the raw data.
}
\examples{
\dontrun{}

}
