% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EKRaw2TSD_oneFile.R
\name{EKRaw2TSD_oneFile}
\alias{EKRaw2TSD_oneFile}
\title{(Internal) Read and Simrad raw file.}
\usage{
EKRaw2TSD_oneFile(i, filelist, prenumt = 10, t = "all", endian = "little",
  timeOffset = 0, minTimeDiff = Inf, msg = TRUE, na.rm = TRUE,
  correctTime = TRUE, bmmd = NULL, TVG.exp = 2, dira_offset = 0,
  cali = TRUE, toTS = FALSE, psze = -7, skipAngles = TRUE,
  cleanNMEA = 1, apply.range.offset = FALSE, thr1m = FALSE)
}
\value{

}
\description{
(Internal) Read and Simrad raw file.
}
\examples{
\dontrun{}

}
