% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EKRaw2TSD_oneFile_write.R,
%   R/compr.event_oneFile_write.R
\name{EKRaw2TSD_oneFile_write}
\alias{EKRaw2TSD_oneFile_write}
\alias{compr.event_oneFile_write}
\title{(Internal) Read and write one Simrad raw file.}
\usage{
EKRaw2TSD_oneFile_write(i, filelist, pingsfiles, vesselfiles, rawvesselfiles,
  beamsfiles, pingsfiles1 = NULL, pingsfiles2 = NULL, pingsfiles3 = NULL,
  prenumt = 10, t = "all", endian = "little", timeOffset = 0,
  minTimeDiff = Inf, msg = TRUE, na.rm = TRUE, correctTime = TRUE,
  bmmd = NULL, TVG.exp = 2, dira_offset = 0, compress = FALSE,
  write = TRUE, cali = TRUE, toTS = FALSE, psze = -7, cleanNMEA = 1,
  tres = NULL, xres = NULL, zres = NULL, rres = NULL, bres = NULL,
  funvbsc = c("median", "mean"), funt = c("median", "mean"),
  adds = list(), split = TRUE, skipAngles = TRUE, origin = 1,
  apply.range.offset = FALSE, thr1m = FALSE, ...)

compr.event_oneFile_write(i, indt, filelist, pingsfiles, vesselfiles,
  beamsfiles, t = "all", drop = TRUE, compress = FALSE, TIME = NULL,
  write = TRUE, tres = NULL, xres = NULL, zres = NULL, rres = NULL,
  bres = NULL, funvbsc = c("mean", "median"), funt = c("median", "mean"),
  adds = list(), split = TRUE, skipAngles = TRUE, origin = 1, z0 = 0,
  pingsnames = NULL, vesselnames = NULL, beamsnames = NULL,
  dumpfiles = NULL, maxlenb = NULL, ...)
}
\value{



}
\description{
(Internal) Read and write one Simrad raw file.

(Internal) Read and write one Simrad raw file.
}
\examples{
\dontrun{}

\dontrun{}

}
