% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.range.TSD.R
\name{extract.range.TSD}
\alias{extract.range.TSD}
\title{Extracts a subset of TSD data according to the ranges given for the variables of 'data' in 'range'.}
\usage{
extract.range.TSD(
  data = list(),
  range = list(),
  this = list(),
  treated = NULL,
  ineq = "<",
  ind.out = FALSE,
  insert.NA = FALSE
)
}
\arguments{
\item{data}{is a list of elements named according to the TSD file format.}

\item{range}{is a list of elements with names matching names if 'data', specifying the range of the corresponding elements.}

\item{this}{is an optional previously generated list of indexes as returned from extract().}

\item{treated}{A vector of variable names already treated.}

\item{ineq}{is a string giving the inequality function to apply to the selection ("<" or "<=").}

\item{ind.out}{is TRUE if the indexes for the elements segmented are to be returned.}

\item{insert.NA}{is TRUE if the discarded data are to be kept as NA.}
}
\value{

}
\description{
Extracts a subset of TSD data according to the ranges given for the variables of 'data' in 'range'.
}
\examples{
\dontrun{}

}
