% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_event.R
\name{extract_event}
\alias{extract_event}
\title{Extracts events from a directory of raw files.}
\usage{
extract_event(event, rawevent, t, ow = TRUE, dir.data = NULL,
  toTSD = TRUE, last = TRUE, temp = FALSE, ...)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{rawevent}{specifies a directory with raw files which are copied to 'event' and converted to TSD filess before running the segmentation.}

\item{t}{is a two element vector of ftim values between which rawfiles are extracted from a directory (see the description of 'event').}

\item{ow}{is FALSE to not overwrite existing data when copying data from 'rawevent'.}

\item{dir.data}{is the path to the directory in which the projects are stored, defaulted by the variable Acoustics_datasets_directory().}

\item{toTSD}{is false to disable converting to TSD files.}

\item{last}{is FALSE to skip the last raw file, useful if this file is not complete.}

\item{temp}{is TRUE to only return the temporary directory if raw files already exist and toTSD==TRUE, in which case the existing and temporary tsd files are not merged into the existing.}

\item{...___}{are parameters passed to EKRaw2TSD(). NA}
}
\value{

}
\description{
Extracts events from a directory of raw files.
}
\examples{
\dontrun{}

}
