% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBeams.R
\name{extractBeams}
\alias{extractBeams}
\title{Extracts a beams from TSD-data, so that if a one dimensional variable has two dimensions, the first is regarded as beams and the apropriate time beams are extracted.}
\usage{
extractBeams(data, beams = 1, var = "all", drop = FALSE)
}
\arguments{
\item{data}{is the list of TSD inputs as returned from read.TSD.}

\item{beams}{is an integer vectir giving the beams to extract.}

\item{var}{is a vector of the TSD-names og the variable to extract the time step from. Variables that are not present in 'var' are returned unaltered. The default var='all' select all variables.}

\item{drop}{Logical: If TRUE drop empty dimensionsof the output.}
}
\description{
Extracts a beams from TSD-data, so that if a one dimensional variable has two dimensions, the first is regarded as beams and the apropriate time beams are extracted.
}
\examples{
\dontrun{}

}
