% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTimeStep.R
\name{extractTimeStep}
\alias{extractTimeStep}
\title{Extracts a time step from TSD-data, so that if a one dimensional variable has two dimensions, the second is regarded as time and the apropriate time steps i extracted.}
\usage{
extractTimeStep(data, t = 1, var = "all")
}
\arguments{
\item{data}{is the list of TSD inputs as returned from read.TSD.}

\item{t}{is a single integer giving the time step to extract.}

\item{var}{is a vector of the TSD-names og the variable to extract the time step from. Variables that are not present in 'var' are returned unaltered. The default var='all' select all variables.}
}
\value{

}
\description{
Extracts a time step from TSD-data, so that if a one dimensional variable has two dimensions, the second is regarded as time and the apropriate time steps i extracted.
}
\examples{
\dontrun{}

}
