% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.matrix.R
\name{extrapolate.matrix}
\alias{extrapolate.matrix}
\title{Extrapolates or shrinks a matrix to according to the specifications given by 'a', 'b' and 'along'. If 'along' (defaulted to 0) is not in 1:nrow(x), 'a' and 'b' are interpreted as the number of rows to add at the beginning and at the end of the columns of the matrix (negative values imply shrinkage). If along is in in 1:nrow(x), 'a' and 'b' are interpreted as the desired lower and upper value to which the values of x[,along] are to be extrapolated by a simple linear expansion of the two first and the two last rows of 'x'. The column given by 'along' should be sorted.}
\usage{
extrapolate.matrix(x, a = NULL, b = NULL, along = 0, grid = FALSE,
  byrow = FALSE, drop.out = TRUE)
}
\arguments{
\item{x}{is the input vector or matrix/data.frame. If 'x' is a vector, it is interpreted as a matrix/data.frame of one column}

\item{a, b}{give the range to which the column given by 'along' is extrapolated/shrinked. If along is not in 1:nrow(x), 'a' and 'b' are interpreted as the number of rows to add at the beginning and at the end of the columns (negative values imply shrinkage). If b==NULL and 'a' has length > 1, the second element of 'a' is interpreted as 'b'.}

\item{along}{is the column along which the extrapolation is specified.}

\item{grid}{is TRUE if the first and the last row of the output are on the grid defined by the expansion.}

\item{byrow}{is TRUE if the expansion/shrinkage is to be done on the rows instead of the columns.}

\item{drop.out}{is TRUE if the output should be stripped of empty dimensions by the drop() function.}
}
\value{

}
\description{
Extrapolates or shrinks a matrix to according to the specifications given by 'a', 'b' and 'along'. If 'along' (defaulted to 0) is not in 1:nrow(x), 'a' and 'b' are interpreted as the number of rows to add at the beginning and at the end of the columns of the matrix (negative values imply shrinkage). If along is in in 1:nrow(x), 'a' and 'b' are interpreted as the desired lower and upper value to which the values of x[,along] are to be extrapolated by a simple linear expansion of the two first and the two last rows of 'x'. The column given by 'along' should be sorted.
}
\examples{
\dontrun{}

}
