% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Expands the fucntion filter() to the edges, instead of returning NA.}
\usage{
filter2(x, filter)
}
\arguments{
\item{x}{is the data vector.}

\item{filter}{is a vector of the filter values as input to filter().}
}
\description{
Expands the fucntion filter() to the edges, instead of returning NA.
}
\examples{
\dontrun{}

}
