% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_voxels.TSD.R
\name{find_voxels.TSD}
\alias{find_voxels.TSD}
\title{Finds the voxels in which the positions in 'pos' are located. NAs returned for voxels outside of the sonar volume.}
\usage{
find_voxels.TSD(pos, data, arr.ind = FALSE)
}
\arguments{
\item{pos}{is a three column matrix of the r, theta, and phi spherical position of the points to locate to voxels.}

\item{data}{is the list of beam configuration data.}

\item{arr.ind}{is TRUE to return array indices, in which voxels outside of the sonar volume are indicated by zeros in one oner more of the columns.}
}
\description{
Finds the voxels in which the positions in 'pos' are located. NAs returned for voxels outside of the sonar volume.
}
\examples{
\dontrun{}

}
