% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.hins.event.R
\name{get.hins.event}
\alias{get.hins.event}
\title{Estimates high intensity noise in acoustic data by identifying high intensity one dimensional spikes. Note that this function only works on MS70 and is therefore deprecated.}
\usage{
get.hins.event(
  con = NULL,
  event = 33,
  cruise = 2009116,
  t = c("20091115084000", "20091115085000"),
  turns = 114,
  k = 5,
  q = c(1000, 100),
  beta_school = 2e-04,
  ind = list(-(1:100), NULL),
  max.memory = 2e+09,
  ...
)
}
\arguments{
\item{con}{is the connection object or a character string naming the output file.}

\item{event}{is the identifier of the event. Given either as the number of the event, a string contained in the name of the event, or the path of the event directory. More than one event may be given, in a vector.}

\item{cruise}{is the identifier of the cruise. Given either as the specification used by IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{t}{is the identifier of the time points. Given either as a vector of integers between 1 and the number of pings in the event, or a vector of time points given as strings "yyyymmddHHMMSS.FFF" or "HHMMSS.FFF" from which the range of the time points are extracted. If t == "all", all files are read and if t == "none" no action is done. If more than one event is given, 't_bgns' must be given as a list of vectors.}

\item{turns}{is a numeric giving the length of the runs, causing 'turns' pings to be read at each step in the function.}

\item{k}{is a numeric giving the width of the medians across pings, used in runmed() inside get.hins.event_small().}

\item{q}{is either a single numeric, or a vector of length 2, givin the start and end point in the linear function along beams defining the threshold times the median filtered values in each direction above which data are classified as high intensity noise.}

\item{beta_school}{is a single numeric representing a typical high school value. Only values above 'beta_school' can be classified as high intensity noise. 'beta_school' assures that spikes due to for example fish in very silent regions of the data are not classified as high intensity noise.}

\item{ind}{is a list of indexes, as given to subset_TSD(), used to select the subset over which the estimation of high intensity noise is done. Defaulted to exclude the first 100 voxels along each beam.}

\item{max.memory}{puts a restiction on the mempry occupied by the function, prompting a call to the user to approve is exceeded.}

\item{...}{variables used in write.TSD() (such as 'ow' for overwriting existing files).}
}
\description{
Estimates high intensity noise in acoustic data by identifying high intensity one dimensional spikes. Note that this function only works on MS70 and is therefore deprecated.
}
\examples{
\dontrun{}

}
