% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nrns.event.R
\name{get.nrns.event}
\alias{get.nrns.event}
\title{Estimates the mean over time steps of close range noise for each voxel of an underwater acoustic system, using mean() or meanexp.quantile() depending on whether the periodic noise is in effect, and writes the result to file in the TSD format if 'con' is given. The near range noise estimates are smoothed with a kernel smoother given by 'filter'. The variance of the Nadaraya-Watson estimator can be derived straight forward by
  var( sum( w_i * x_i ) )
<- sum( var( w_i * x_i ) )
<- sum( w_i^2 * var( x_i ) )
<- sum( w_i^2 * s^2 )
<- sum( w_i^2 ) * s^2 
<- sum( (k_i / sum(k_i))^2 ) * s^2 
<- sum( k_i^2 / sum(k_i)^2 ) * s^2 
<- sum( k_i^2 ) / sum(k_i)^2 * s^2}
\usage{
get.nrns.event(
  con = NULL,
  event = NULL,
  cruise = NULL,
  t = NULL,
  bgns = NULL,
  filter = dnorm(-20:20, sd = 4),
  nsdn = 1000,
  farv = 1000,
  surface = 451:500,
  esnm = "MS70",
  ...
)
}
\arguments{
\item{con}{is the connection object or a character string naming the output file.}

\item{event}{is the identifier of the event. Given either as the number of the event, a string contained in the name of the event, or the path of the event directory. More than one event may be given, in a vector.}

\item{cruise}{is the identifier of the cruise. Given either as the specification used by IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{t}{is the identifier of the time points. Given either as a vector of integers between 1 and the number of pings in the event, or a vector of time points given as strings "yyyymmddHHMMSS.FFF" or "HHMMSS.FFF" from which the range of the time points are extracted. If t=="all", all files are read and if t=="none" no action is done. If more than one event is given, 't_bgns' must be given as a list of vectors.}

\item{bgns}{is either a list containing the vector of background noise 'bgns' of each beam, og the vector itsself.}

\item{filter}{is the kernel used when smoothing the near range noise of each beam.}

\item{nsdn}{is the number of standard deviation above the mean nrea range noise for each beam, defining the values that are set to 0. The beam is traced from the sonar and out, and at the first value below the threshold, the near range noise is set to zero for all values at and beyond this value.}

\item{farv}{is the number of far voxels used for the calclulation of the mean and standard deviation used when assigning zeros to the smoothed near range noise. For farv=13, the 13 farthest non-missing values are used for each beam.}

\item{surface}{is a vector of indexes for the beams affected by the surface noise (extrapolated in the near range noise estimation).}

\item{esnm}{Deprecated, use full path instead. Specifies the acoustical instrument to use.}

\item{...}{variables used in write.TSD() (such as 'ow' for overwriting existing files).}
}
\description{
Estimates the mean over time steps of close range noise for each voxel of an underwater acoustic system, using mean() or meanexp.quantile() depending on whether the periodic noise is in effect, and writes the result to file in the TSD format if 'con' is given. The near range noise estimates are smoothed with a kernel smoother given by 'filter'. The variance of the Nadaraya-Watson estimator can be derived straight forward by
  var( sum( w_i * x_i ) )
<- sum( var( w_i * x_i ) )
<- sum( w_i^2 * var( x_i ) )
<- sum( w_i^2 * s^2 )
<- sum( w_i^2 ) * s^2 
<- sum( (k_i / sum(k_i))^2 ) * s^2 
<- sum( k_i^2 / sum(k_i)^2 ) * s^2 
<- sum( k_i^2 ) / sum(k_i)^2 * s^2
}
\examples{
\dontrun{}

}
