% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pdns_phase.TSD.R
\name{get.pdns_phase.TSD}
\alias{get.pdns_phase.TSD}
\title{Estimates the phase of the periodic noise of the MS70 sonar. If the variables 'badb', 'pns1', 'pns2', 'harm' and 'bgns' are not present in the input list 'data', an attempt is made to read them from the directory holding the MS70 noise estimate for the given event.}
\usage{
get.pdns_phase.TSD(data = list(), nsind = 0.75, cruise = 2009116,
  event = NULL, esnm = "MS70", dir.data = NULL, pdns_scale = 1e-14,
  TVG = TRUE)
}
\arguments{
\item{data}{is a list of the following variables ( can be obtained by data=c(read.event(var=c("beams","vbsc","")),read.TSDs(noise.path(),var=c("pns1","pns2","harm","badb","bgns","acfq"))) ):}

\item{nsind}{is a vector of indexes along the beams, as input to ind.expand(), used to select the subset over which the estimation of the phase of the periodic noise is done. If given as a single numeric, the outermost 'nsind' voxels are used in each beam.}

\item{cruise, event, esnm}{Used in \code{\link{noise.path}} to get the path to the noise data holding the background noise, if missing in the data.}

\item{dir.data}{is the path to the directory in which the projects are stored, defaulted by the variable Acoustics_datasets_directory().}

\item{pdns_scale}{is used to scale the noise in order to allow the optimization to work.}

\item{TVG}{is FALSE if TVG compensation is to be removed from the data.}
}
\value{

}
\description{
Estimates the phase of the periodic noise of the MS70 sonar. If the variables 'badb', 'pns1', 'pns2', 'harm' and 'bgns' are not present in the input list 'data', an attempt is made to read them from the directory holding the MS70 noise estimate for the given event.
}
\examples{
\dontrun{}

}
