% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRuns.R
\name{getRuns}
\alias{getRuns}
\title{Generates runs of length 'n' of the vector 't' of time steps.}
\usage{
getRuns(n, t)
}
\arguments{
\item{n}{is the length of the runs.}

\item{t}{is a vector of time steps to group into runs.}
}
\description{
Generates runs of length 'n' of the vector 't' of time steps.
}
\examples{
\dontrun{}

}
