% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVoxelWidth2D.R
\name{getVoxelWidth2D}
\alias{getVoxelWidth2D}
\title{Calculates edge values for voxels in a 2-D sonar fan. If a vertically oriented fan is}
\usage{
getVoxelWidth2D(data, fanWidth = "b2")
}
\arguments{
\item{data}{is the list of the inputs used in the function as returned from read.TSD() (must include "").}

\item{fanWidth}{has a number of possible values: (1) "b1": one way beam width. (2) "b2": two way beam width. (3) "fe": beams modeled by rectangular cones with width withing the fan given by the inter-beam angle, and calculated using the equivalent beam angle. This normally causes larged fan width due to overlap between beams.}
}
\description{
Calculates edge values for voxels in a 2-D sonar fan. If a vertically oriented fan is
}
\examples{
\dontrun{}

}
