% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getzfromctd.R
\name{getzfromctd}
\alias{getzfromctd}
\title{Computes the z-position based on CTD data (if missing).}
\usage{
getzfromctd(ctd, Pain = FALSE)
}
\arguments{
\item{ctd}{is the conductivity, temperature and depth data.}

\item{Pain}{is TRUE if pressure "ihpr" is given in Pascal and FALSE if given in decibar relative to surface pressure (10000 Pascal, giving values approximately equivalent to water depth).}
}
\description{
Computes the z-position based on CTD data (if missing).
}
\examples{
\dontrun{}

}
