% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harx.SH80_oneping.R
\name{harx.SH80_oneping}
\alias{harx.SH80_oneping}
\title{Calculates the horizontal area of the voxels of the SH80 multibeam sonar given the CTD-data and beams-data of the particular data set 'dataset'. See read.event() for importing volumes of voxels.}
\usage{
harx.SH80_oneping(data)
}
\arguments{
\item{data}{is the list of the inputs used in the function as returned from read.TSD() (must include "").}
}
\description{
Calculates the horizontal area of the voxels of the SH80 multibeam sonar given the CTD-data and beams-data of the particular data set 'dataset'. See read.event() for importing volumes of voxels.
}
\examples{
\dontrun{}

}
