% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sonar.R
\name{is.sonar}
\alias{is.sonar}
\title{Identifies whether the acoustic system is a sonar, fishery sonar or echosounder, for exactly one ping is 'data' is given.}
\usage{
is.sonar(esnm = "MS70", bydirs = FALSE, data = list(), margin = 10 *
  pi/180, bmmd = NULL, fishery = FALSE)
}
\arguments{
\item{esnm}{the four character identifyer of the acoustical system. Currently implemented: echosounders: "EK60", "ME70"; sonars: "MS70", "SX80", "SH80", "SX90", "SH90".}

\item{bydirs}{is TRUE if the elevation angle of the beams should be used to identify sonars by the requirement that none of the beams should be within 'margin' of the vertical orientation.}

\item{data}{is an optional list of beam configuration data.}

\item{margin}{is the margin around the vertical orientation outside which all beams must be to be identified as sonar.}

\item{bmmd}{is the desired mean mode of the system, which should be set for instance for fishery sonars.}

\item{fishery}{is TRUE to return TRUE only for fishery sonars (not including MS70).}
}
\value{

}
\description{
Identifies whether the acoustic system is a sonar, fishery sonar or echosounder, for exactly one ping is 'data' is given.
}
\examples{
\dontrun{}

}
