% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.files_caseInsensitive.R
\name{list.files_caseInsensitive}
\alias{list.files_caseInsensitive}
\title{Sorts in a case insensitive matter. Created to solve the problem of errors when listing files in R as run from the terminal, in which case sorting is not case insensitive. In the R-GUI sorting is done case insensitive.}
\usage{
list.files_caseInsensitive(path = ".", pattern = NULL,
  all.files = FALSE, full.names = FALSE, recursive = FALSE,
  ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE,
  method = "shell")
}
\arguments{
\item{method}{The ordering method used in \code{\link{order}}.}

\item{...}{Used in \code{\link{list.files}}.}
}
\value{

}
\description{
Sorts in a case insensitive matter. Created to solve the problem of errors when listing files in R as run from the terminal, in which case sorting is not case insensitive. In the R-GUI sorting is done case insensitive.
}
\examples{
\dontrun{}

}
