% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanexp.quantile.R
\name{meanexp.quantile}
\alias{meanexp.quantile}
\title{Estimates the parameter of an exponentially distributed 'X' based on the quantile of values. F(x) = 1 - exp(-x/beta) = p => betahat = quantile(x,p) / log(1/(1-p)). Arrays up to 3 dimensions are accepted.}
\usage{
meanexp.quantile(x, prob = 0.5, type = 6, MARGIN = NULL)
}
\arguments{
\item{x}{is the data.}

\item{prob}{is the probability.}

\item{type}{is the type of the quantile calculation (see quantile()).}

\item{MARGIN}{is used in the same way as in apply().}
}
\value{

}
\description{
Estimates the parameter of an exponentially distributed 'X' based on the quantile of values. F(x) = 1 - exp(-x/beta) = p => betahat = quantile(x,p) / log(1/(1-p)). Arrays up to 3 dimensions are accepted.
}
\examples{
\dontrun{}

}
