% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanexp.quantile_a.R
\name{meanexp.quantile_a}
\alias{meanexp.quantile_a}
\title{Estimates the parameter 'beta' of an exponentially distributed 'X' based on the quantile of values. This function accepts an addition to the expected value of 'X', 'a', which must be known. The calculation is an optimizatoin of 'beta' from an expression deduced from the cummulative distribution function (CDF) of an arbitrary 'X' (which is the mean of the CDF of X_t):
X_t ~ exp(beta_t), where beta_t = beta + a_t
F_X(x) = 1/N * sum_t=1^N(F_X_t(x_t)) = 1 - 1/N * sum_t=1^N(exp(-x/beta_t)),
which is estimated by
u = 1 - 1/N * sum_t=1^N(exp(-Q(u)/beta_t))
u = 1 - 1/N * sum_t=1^N(exp(-Q(u)/(beta+a_t))), where Q(u) is the u-quantile.
This expression is solved numerically for 'beta'}
\usage{
meanexp.quantile_a(x, lower = 0, upper = 1, prob = 0.5, a = 1,
  type = 6, MARGIN = NULL)
}
\arguments{
\item{x}{is the data.}

\item{prob}{is the probability.}

\item{type}{is the type of the quantile calculation (see quantile()).}

\item{MARGIN}{is used in the same way as in apply().}
}
\value{

}
\description{
Estimates the parameter 'beta' of an exponentially distributed 'X' based on the quantile of values. This function accepts an addition to the expected value of 'X', 'a', which must be known. The calculation is an optimizatoin of 'beta' from an expression deduced from the cummulative distribution function (CDF) of an arbitrary 'X' (which is the mean of the CDF of X_t):
X_t ~ exp(beta_t), where beta_t = beta + a_t
F_X(x) = 1/N * sum_t=1^N(F_X_t(x_t)) = 1 - 1/N * sum_t=1^N(exp(-x/beta_t)),
which is estimated by
u = 1 - 1/N * sum_t=1^N(exp(-Q(u)/beta_t))
u = 1 - 1/N * sum_t=1^N(exp(-Q(u)/(beta+a_t))), where Q(u) is the u-quantile.
This expression is solved numerically for 'beta'
}
\examples{
\dontrun{}

}
