% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median3d.R
\name{median3d}
\alias{median3d}
\title{Median along the dimension given by 'along'. The same as apply(x,(1:3)[-along],median,na.rm=TRUE), but some hundred times faster.}
\usage{
median3d(x, along = 1)
}
\arguments{
\item{x}{is a 3d array to be median smoothed along the 'along' dimension.}
}
\value{

}
\description{
Median along the dimension given by 'along'. The same as apply(x,(1:3)[-along],median,na.rm=TRUE), but some hundred times faster.
}
\examples{
\dontrun{}

}
