% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medSmooth1.R
\name{medSmooth1}
\alias{medSmooth1}
\title{Median smoothing an array along the first dimension.}
\usage{
medSmooth1(x, w = 3, try.runmed = FALSE, silent = FALSE, ...)
}
\arguments{
\item{x}{An array to be median smoothed along the first dimension.}

\item{w}{The width of median filter (odd number).}

\item{try.runmed}{Locigal: If TRUE try the function \code{\link{medSmooth1_runmed}} first and then \code{\link{medSmooth1_median3d}}.}

\item{silent}{Used in \code{\link{medSmooth1_runmed}}.}

\item{...}{Used in \code{\link{medSmooth1_runmed}} and \code{\link{medSmooth1_median3d}}.}
}
\value{

}
\description{
Median smoothing an array along the first dimension.
}
\examples{
\dontrun{}

}
