% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medSmooth1_median3d.R
\name{medSmooth1_median3d}
\alias{medSmooth1_median3d}
\title{Median smoothing an array along the first dimension.}
\usage{
medSmooth1_median3d(x, w = 3, maxsize = 2e+08)
}
\arguments{
\item{x}{is an array to be median smoothed along the first dimension.}

\item{w}{is the width of median filter (odd number).}

\item{maxsize}{is the maximum size of the array used in the median smoothing with median3d(). If prod(dim(x)[1]+w-1,dim(x)[-1],w)>maxsize, an error is issued.}
}
\value{

}
\description{
Median smoothing an array along the first dimension.
}
\examples{
\dontrun{}

}
