% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medSmooth1_runmed.R
\name{medSmooth1_runmed}
\alias{medSmooth1_runmed}
\title{Median smoothing an array along the first dimension.}
\usage{
medSmooth1_runmed(x, w = 3, ...)
}
\arguments{
\item{x}{is an array to be median smoothed along the first dimension.}

\item{w}{is the width of median filter (odd number).}

\item{...}{Used for robustness.}
}
\value{

}
\description{
Median smoothing an array along the first dimension.
}
\examples{
\dontrun{}

}
