% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_events.R
\name{merge_events}
\alias{merge_events}
\title{(Internal) Merge segmentation events, used in SX90.segment.event() and extract_event().}
\usage{
merge_events(events, event, cruise = NULL, esnm = NULL, dir.data = NULL,
  ctd = 1)
}
\value{

}
\description{
(Internal) Merge segmentation events, used in SX90.segment.event() and extract_event().
}
\examples{
\dontrun{}

}
