% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_TSD_by_Time.R
\name{merge_TSD_by_Time}
\alias{merge_TSD_by_Time}
\title{Merges the last file of the current file number and the first file of the next.}
\usage{
merge_TSD_by_Time(files, fun = weighted.mean, msg = FALSE, ...)
}
\arguments{
\item{files}{list of file names, seed the use in \code{\link{compr.event}}.}

\item{fun}{funciton specifying how to merge the information in the common time interval.}

\item{msg}{Logical: if TRUE print status bar.}

\item{...}{Passed on to \code{fun} and \code{\link{write.TSD}}.}
}
\value{

}
\description{
Merges the last file of the current file number and the first file of the next.
}
\examples{
\dontrun{}

}
