% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_TSD_by_var.R
\name{merge_TSD_by_var}
\alias{merge_TSD_by_var}
\title{This function returns a list of indices of files of common time (utim)/sailed distance (sadv), or if this list is given merges the common files.}
\usage{
merge_TSD_by_var(
  files,
  var = "utim",
  ind = NULL,
  fun = weighted.mean,
  msg = FALSE,
  ...
)
}
\arguments{
\item{files}{list of file names, seed the use in \code{\link{compr.event}}.}

\item{var}{The variable used as the merging variable, e.g., "utim" or "sadv".}

\item{ind}{A list of indices of common files. Each element of the list contains two or more file indices of files to be merged.}

\item{fun}{funciton specifying how to merge the information in the common time interval.}

\item{msg}{Logical: if TRUE print status bar.}

\item{...}{Passed on to \code{fun} and \code{\link{write.TSD}}.}
}
\value{

}
\description{
This function returns a list of indices of files of common time (utim)/sailed distance (sadv), or if this list is given merges the common files.
}
\examples{
\dontrun{}

}
