% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.path.R
\name{noise.path}
\alias{noise.path}
\title{Locates the paths to files holding the noise estimates and the estimates of the correlation between voxels. First the relevant files are seached for in the event given by 'event', 'cruise', and 'esnm', and then the files are seached for in the echoIBM/Resources/Noise/Main directory by matching 'esnm' and 'utim', or by 'esnm', 'cruise' if 'utim' is not given. This may select more than one relevant file, and read.TSDs() should be applied to the detected files, possibly with clean=FALSE.}
\usage{
noise.path(
  event = NULL,
  cruise = 2009116,
  esnm = "MS70",
  utim = NULL,
  dir.data = NULL,
  noisevar = c("bgns", "nrnp"),
  corvar = c("crb1", "olpb")
)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{cruise}{is either the idenfication number of the cruise (or to the directory "/Applications/echoIBM/Resources" in which default noise files for echoIBM are located), given as specified by the IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{utim}{is the unix time by which the noise files should be selected.}

\item{dir.data}{is the path to the directory in which the projects are stored, defaulted by the variable Acoustics_datasets_directory().}

\item{noisevar}{is a vetor of TSD names which mest be present in the noise estimate files.}

\item{corvar}{is a vetor of TSD names which mest be present in the correlation estimate files.}
}
\description{
Locates the paths to files holding the noise estimates and the estimates of the correlation between voxels. First the relevant files are seached for in the event given by 'event', 'cruise', and 'esnm', and then the files are seached for in the echoIBM/Resources/Noise/Main directory by matching 'esnm' and 'utim', or by 'esnm', 'cruise' if 'utim' is not given. This may select more than one relevant file, and read.TSDs() should be applied to the detected files, possibly with clean=FALSE.
}
\examples{
\dontrun{}

}
