% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol.school.R
\name{pol.school}
\alias{pol.school}
\title{Cacluates the polarization of a school as one of the following methods: (1) The mean angle deviation between the individual directions and the mean direction of the school, given in radians (Huth and Wissel 1992, Parrish et al 2002). (2) The norm of the average direction unit vectors (Couzin et al 2002). (3) The simple standard deviation of the azimuth and elevation angles of the fish (Holmin et al 2011 (Seattle meeting)).}
\usage{
pol.school(x, type = c("Huth", "Couzin", "sd", "azimuth", "elevation"))
}
\arguments{
\item{x}{is either a list of elements 'vlxf', 'vlyf' and 'vlzf', or 'rtzf' and 'rtxf', or a matrix of vectors along the rows.}

\item{type}{is a string representing the type of polarization estimate. Currently implemented are: (1) The mean angle deviation between the individual directions and the mean direction of the school (Huth and Wissel 1992, Parrish et al 2002). (2) The norm of the average direction unit vectors (Couzin et al 2002). (3) The simple standard deviation of the azimuth and elevation angles of the fish (Holmin et al 2011 (Seattle meeting)).}
}
\value{

}
\description{
Cacluates the polarization of a school as one of the following methods: (1) The mean angle deviation between the individual directions and the mean direction of the school, given in radians (Huth and Wissel 1992, Parrish et al 2002). (2) The norm of the average direction unit vectors (Couzin et al 2002). (3) The simple standard deviation of the azimuth and elevation angles of the fish (Holmin et al 2011 (Seattle meeting)).
}
\examples{
\dontrun{}

}
