% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psx.TSD.R
\name{psx.TSD}
\alias{psx.TSD}
\title{(Internal) Get midpoints of voxels.}
\usage{
psx.TSD(x, pad = TRUE, split = TRUE, ...)
}
\value{

}
\description{
(Internal) Get midpoints of voxels.
}
\examples{
\dontrun{}

}
