% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.event_insertNA.R
\name{read.event_insertNA}
\alias{read.event_insertNA}
\title{Adds empty list elements of NAs (if vectors) to data read in read.event(). This is done by comparing to the time step indices stored in the object 'TIME'.}
\usage{
read.event_insertNA(thisout, TIME, filesind, filelist, t, tlist, var)
}
\arguments{
\item{thisout}{is a list containing the relevant data.}

\item{TIME}{is the list returned from UNIX_time().}

\item{filesind}{is a vector of the indexes of the relevant files in the list of files.}

\item{filelist}{is the list of files.}

\item{t}{is the tine step indexes to return data at.}

\item{tlist}{is the list of time step indexes for the files.}

\item{var}{is a vector of the variables to to insert NAs to.}
}
\value{

}
\description{
Adds empty list elements of NAs (if vectors) to data read in read.event(). This is done by comparing to the time step indices stored in the object 'TIME'.
}
\examples{
\dontrun{}

}
