% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.event_read_other_ctd.R
\name{read.event_read_other_ctd}
\alias{read.event_read_other_ctd}
\title{Reads ctd-variables. Used in read.event().}
\usage{
read.event_read_other_ctd(var, ctdfiles, ctdfilesind, TIME)
}
\arguments{
\item{var}{is a vector of the variables to read.}

\item{ctdfiles}{is a vector of the file names of the ctd-files.}

\item{ctdfilesind}{is a vector of indices of the ctd-files in the list of files associated with 'TIME'.}

\item{TIME}{is the list returned from UNIX_time().}
}
\description{
Reads ctd-variables. Used in read.event().
}
\examples{
\dontrun{}

}
