% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.event_sgsc2pos.TSD.R
\name{read.event_sgsc2pos.TSD}
\alias{read.event_sgsc2pos.TSD}
\title{Transforms from 'sgsc' to positions, in a similar way as used in pplot3d.TSD(). Used in read.event() for clustering 'sgsc'.}
\usage{
read.event_sgsc2pos.TSD(data, N = 100, esnm = "MS70", ideal = TRUE,
  allert = NULL, seabed = -12000, rot = 1, compensation = c("pitch",
  "roll"), ...)
}
\arguments{
\item{data}{is the list of TSD inputs as returned from read.TSD(var=c("vbsc","voxels","vessel")). Should only contain one ping/timestep.}

\item{N}{is the approximate number of points plotted.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{ideal}{is TRUE to represent the simple case where the speed of sound 'data$asps' is invariant of depth.}

\item{allert}{is the limit of the number of points inducing a warning asking the user to continue. Default is NULL, indicating no check for the number of points to be generated.}

\item{seabed}{is the depth of the seabed, at which the beams are reflected when calculating the midpoints of voxels.}

\item{rot}{see soundbeam.TSD().}

\item{compensation}{is a vector of string giving which rotation values that are compensated for in the sonar. Only c("pitch","roll") is available for the current version. Used in soundbeam.TSD.}
}
\value{

}
\description{
Transforms from 'sgsc' to positions, in a similar way as used in pplot3d.TSD(). Used in read.event() for clustering 'sgsc'.
}
\examples{
\dontrun{}





}
