% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.event_unzip_vbsc.R
\name{read.event_unzip_vbsc}
\alias{read.event_unzip_vbsc}
\title{Merges time steps of acoustic data located in the neighbor directory "temp" of the "tsd"-directory given by 'event'.}
\usage{
read.event_unzip_vbsc(x, pad = TRUE, split = TRUE, filename = NULL,
  t = NULL, fill = NA, keep.vxIX = TRUE)
}
\arguments{
\item{x}{a list containing the possibly compressed acoustic data (vbsc, vxIX).}

\item{pad}{is TRUE to pad with missing values to make the list elements fit in the dimension up to the last dimension.}

\item{split}{Logical: if TRUE, split each list element to have last dimenstion equal to 1.}

\item{filename}{is the filename of the file containing the data, used in warning messages.}

\item{t}{is the time index of the data read, also used in warning messages.}

\item{fill}{is the value to use for missing data (outside of the mask).}

\item{keep.vxIX}{is TRUE to return also the indices at the non-missing acoustic data.}
}
\description{
Merges time steps of acoustic data located in the neighbor directory "temp" of the "tsd"-directory given by 'event'.
}
