% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfosPP.R, R/readProfosPP.R
\name{plotProfosPP}
\alias{plotProfosPP}
\alias{readProfosPP}
\title{Plot output from readProfosPP.}
\usage{
plotProfosPP(
  x,
  ind = c(1, 2),
  col = list(ship = c("black", "purple"), school = c("black", "brown", "orange"), link =
    "green4"),
  pch = list(ship = "*", school = 1),
  cex = list(ship = c(1, 0.5), school = c(1, 0.5, 0.5), link = 1),
  lwd = list(track = 0.5, link = 1, sv = 1),
  plotlink = TRUE,
  linkvar = "lowess",
  every = NULL,
  use = c("sv", "Sv"),
  svQ = c(0.1, 0.9),
  ...
)

readProfosPP(
  x,
  currentSpeed = NULL,
  currentAngle = NULL,
  compensateCurrent = TRUE,
  sparfact = 1,
  spar0 = 0.5,
  byId = TRUE,
  rm.duplicates = FALSE,
  filter = NULL,
  cores = 1,
  nrows = -1,
  smooth = TRUE,
  ...
)
}
\arguments{
\item{x}{The Profos per ping file.}

\item{ind}{A numeric vector indicating which plots to produce, 1 being the track plot and 2 the sv plot.}

\item{col, pch, cex}{A list of color, ploting character and character expansion values for the ship, school and links between these at transitions between weak and strong echo.}

\item{lwd}{A list of line width values for the (smoothed) tracks (plotted with type="o"), the links, and the sv.}

\item{plotlink}{Logical: If TRUE plot links.}

\item{linkvar}{The type of smoothing of the school to connect ship positions to, one of "spline" (less smoothing) and "lowess" (stronger smoothing).}

\item{every}{Plot links at every \code{every} ping.}

\item{use}{Whether to use linear ('sv') or logarithmical ('Sv') mean volume backscattering.}

\item{svQ}{A vector of length 2 defining the quantiles used for defining the amplitude of the sine wave which is fitted to the mean Sv to obtain transitions between weak and strong echo.}

\item{...}{Arguments passed to \code{\link{lowess}}.}

\item{currentSpeed}{A vector of the speed of the current at the observation, one value per ping or one value for all pings.}

\item{currentAngle}{A vector of the angle of the current at the observation, in degrees clockwise from North, one value per ping or one value for all pings.}

\item{compensateCurrent}{Logical: If TRUE (the default), compensate for the current.}

\item{sparfact}{A factor to multiply the \code{spar} used in \code{\link{smooth.spline}} when smoothing the school positions.}

\item{spar0}{The value of \code{spar} used if application of the \code{sparfact} results in negative values for \code{spar}.}
}
\value{



}
\description{
This function creates two plots, (1) one showing the raw and spline smoothed track of the vessel (ship) and the raw and spline and lowess smoothed track of the school, and (2) a related plot showing the mean volume backscattering coefficient/strength of the school as a function of incidence angle rotation (increasing along the unit circle).
}
\examples{
\dontrun{}

\dontrun{}

}
