% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDuplicated_tlist.R
\name{removeDuplicated_tlist}
\alias{removeDuplicated_tlist}
\title{Removes duplicated time steps from the list of time step indices 'tlist'.}
\usage{
removeDuplicated_tlist(tlist, files, ind = NULL)
}
\arguments{
\item{tlist}{is a list of time steps for each file in the vector of file paths 'files'.}

\item{files}{is a list of file names.}

\item{ind}{is a vector of indices for the files for which to check for duplicated time steps.}
}
\description{
Removes duplicated time steps from the list of time step indices 'tlist'.
}
\examples{
\dontrun{}

}
