% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleHigh.R
\name{resampleHigh}
\alias{resampleHigh}
\title{Resamples large values from the remaining values (useful in analyses containing suspucously high values which desrupt the analysis).}
\usage{
resampleHigh(x, high, trim = 0.1)
}
\arguments{
\item{x}{is the vector to modify.}

\item{high}{is the boundary classifying values as high values, defined by x/mean(x,trim=trim) (values larger than 'high' times the trimmed mean are defined as high values).}

\item{trim}{is used in x/mean(x,trim=trim).}
}
\description{
Resamples large values from the remaining values (useful in analyses containing suspucously high values which desrupt the analysis).
}
\examples{
\dontrun{}

}
