% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmUnix_Sgpm.R
\name{rmUnix_Sgpm}
\alias{rmUnix_Sgpm}
\title{Remove the files with basenames UNIX_time.tsd and sgPM.tsd. Used in UNIX_time() and read.event().}
\usage{
rmUnix_Sgpm(x)
}
\arguments{
\item{x}{A vector of file names.}
}
\description{
Remove the files with basenames UNIX_time.tsd and sgPM.tsd. Used in UNIX_time() and read.event().
}
