% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotfun.R
\name{rotfun}
\alias{rotfun}
\title{Returns a function representing the product of three dimensional rotation matrices, specified by the string 'ax'.}
\usage{
rotfun(ax)
}
\arguments{
\item{ax}{is a string containing a sequence of the characters "x", "y" and "z".}
}
\description{
Returns a function representing the product of three dimensional rotation matrices, specified by the string 'ax'.
}
\examples{
\dontrun{}

}
