% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotz.R
\name{rotz}
\alias{rotz}
\title{Returns three dimensional rotation matrices for rotation around the z-axis of a three dimensional coordinate system.}
\usage{
rotz(ang, radians = TRUE, block.out = FALSE, drop.out = TRUE, byrow = TRUE)
}
\arguments{
\item{ang}{is a vedctor of the rotation angles around the z-axis.}

\item{radians}{is TRUE if 'ang' is given in radians, and FALSE if 'ang' is given in degrees.}

\item{block.out}{is TRUE if the rotation matrices should be returned as a 3x3*l matrix, where 'l' is the number of rotation angles. If FALSE, an 3x3xl array is returned.}

\item{drop.out}{is relevant only if block.out==FALSE and l==1, and if set to TRUE a 3x3 matrix is returned instead of a 3x3x1 array.}

\item{byrow}{is TRUE if the matrices are to be organized by rows (vertically).}
}
\value{

}
\description{
Returns three dimensional rotation matrices for rotation around the z-axis of a three dimensional coordinate system.
}
\examples{
\dontrun{}

}
