% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmaskEchosounder.R
\name{segmaskEchosounder}
\alias{segmaskEchosounder}
\title{Create theoretical segmentation data for echosounder.}
\usage{
segmaskEchosounder(event, t = "all")
}
\arguments{
\item{event}{The path to the event holding TSD files.}

\item{t}{The time steps to generate theoretical segmentation for.}
}
\description{
Create theoretical segmentation data for echosounder.
}
