% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simrad_bwt.R
\name{Simrad_bwt}
\alias{Simrad_bwt}
\title{Converts from Simrad beam widths to x and y beam widths.}
\usage{
Simrad_bwt(data)
}
\arguments{
\item{data}{is a list containing the required variables 'dirx' (vector of x-beam-directions), 'diry' (vector of y-beam-directions), 'dirz' (vector of z-beam-directions), and 'esnm' (name of the echosounder/sonar).}
}
\description{
Converts from Simrad beam widths to x and y beam widths.
}
\examples{
\dontrun{}

}
