% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simrad_dir.R
\name{Simrad_dir}
\alias{Simrad_dir}
\title{Converts from Simrad beam directions to azimuth and elevation angles.}
\usage{
Simrad_dir(data, dira_offset = 0)
}
\arguments{
\item{data}{is a list containing the rquired variables 'dirx' (vector of x-beam-directions), 'diry' (vector of y-beam-directions), 'dirz' (vector of z-beam-directions), and 'esnm' (name of the echosounder/sonar).}

\item{dira_offset}{An optional offset in the azimuth direction 'dira'.}
}
\description{
Converts from Simrad beam directions to azimuth and elevation angles.
}
\examples{
\dontrun{}

}
