% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{sonR}
\alias{sonR}
\alias{sonR-package}
\title{Library for reading, processing and converting sonar data to the TSD format}
\description{
This package contains many functions used by the packages echoIBM and cpplot3d.
}
\details{
Details of the package
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/arnejohannesholmin/sonR/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@imr.no}

}
