% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonR_implemented.R
\name{sonR_implemented}
\alias{sonR_implemented}
\title{Test for implemented acosutical systems in echoIBM().}
\usage{
sonR_implemented(esnm, type = c("MBS", "MBE", "SBE", "OFS"), notest = FALSE)
}
\arguments{
\item{esnm}{is the name of the acoustical instrument, given as a four character string, or a list containing an element named "esnm". May be given in lower case.}

\item{type}{is a vector of the types of acoustical instrument to check for implementation in echoIBM. As default the function tests against all implemeneted systems.}

\item{notest}{is TRUE to return the implemented systems, in which case no test is made.}
}
\value{

}
\description{
Test for implemented acosutical systems in echoIBM().
}
\examples{
\dontrun{}

}
