% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundbeam.R
\name{soundbeam}
\alias{soundbeam}
\title{Computes the path of a ray of sound for the given temperature, salinity and depth/pressure values. The sound beam is calculated in the x-y-plane, where y corresponds to z in three dimensions. If unequal in length, the shorter ones of the inputs are recycled. A list of the following elements is returned Used in soundbeam.TSD().}
\usage{
soundbeam(
  start = c(0, 0),
  ang,
  ctd,
  seabed = -12000,
  time = NULL,
  w = NULL,
  lenb = 1500,
  sint = 0.000512,
  rpos = c("midpoint", "edge"),
  Ncirc = 1000,
  maxlength = Inf,
  plot = TRUE,
  Pain = TRUE
)
}
\arguments{
\item{start}{is the transducer position.}

\item{ang}{is a matrix of two columns representing the vectors of the start and end angles of the circle segments along which the sound beam is traced for each sound speed layer, as defined on the unit circle.}

\item{ctd}{is the conductivity-temperature-depth data given as a list of suitable names.}

\item{seabed}{is the z-position of the sea bed.}

\item{time}{is the vector of time points at which the positions along the sound beam are recorded. If not given, 'time' is calculated from 'lenb' and 'sint'.}

\item{w}{is the vector of positions along the sound beam, overriding 'time' or 'lenb' and 'sint'.}

\item{lenb}{is the number of partitions radially of equal distance 'sint' in time.}

\item{sint}{is the time length of the sound pulse. 'lenb' and 'sint' are used specifically when calculating midpoints of voxels of sonars, where the sound beam travels out to an object and is reflected back to the sonar. Thus the effective time used is dt=sint/2 (see the master thesis of Arne Johannes Holmin page 30-33). The first time point starts at dt/4, and the remaining time points are dt*k, k=2,3,...,lenb. (See the master thesis of Arne Johannes Holmin page 31-32.)}

\item{rpos}{is a string of two possible values: "midpoint" which specifies that the positions alont the beams are to bo located at midpoints of voxels and "edge" which returns the edges of the voxels including the first edge on the surface of the sonar (length of output is lenb+1 in this case). Abbreviation are allowed.}

\item{Ncirc}{are the maximum number of circles from which the soundbeams are reconstructed.}

\item{maxlength}{is the maximum length of the beam, restricting the number of calculations of circles.}

\item{plot}{is TRUE if the sound beams should be plotted (time consuming).}

\item{x}{is a vector of the x-positions of the sound beam.}

\item{y}{is a vector of the y-positions of the sound beam.}

\item{l}{is a vector of the tracked positions along the sound beam.}

\item{r}{is a vector of the origins of the circles along which the sound beam is traced for each sound speed layer.}

\item{origin}{is a matrix of two columns representing the x- and y-positions of the points linking the circle segments.}

\item{pos}{is a vector of the x-positions of the sound beam.}

\item{extrapolated}{is a logical vector which is TRUE for all circle segments located in sound speed layers extrapolated from the ctd-data, in the case that the ctd measurements do not extend the entire water column.}

\item{totlength}{is the total length of the calculated path along which the sound beam positions are tracked.}
}
\description{
Computes the path of a ray of sound for the given temperature, salinity and depth/pressure values. The sound beam is calculated in the x-y-plane, where y corresponds to z in three dimensions. If unequal in length, the shorter ones of the inputs are recycled. A list of the following elements is returned Used in soundbeam.TSD().
}
\examples{
\dontrun{}

}
