% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedofsound.R
\name{speedofsound}
\alias{speedofsound}
\title{Computes the speed of sound of the ocean for the given temperature, salinity and pressure values. Method adopted from "sw_svel.m" and based on UNESCO 1983.}
\usage{
speedofsound(T, S, P, Z = NULL, P0 = 10.1325, rho = 1027, g = 9.82,
  Pain = TRUE)
}
\arguments{
\item{T}{is the temperature in degrees celcius, or a list of the inputs in the TSD file format.}

\item{S}{are the salinity values in parts per thousand.}

\item{P}{are the pressure values in decibar or Pascal.}

\item{Z}{are the z-positions in meters (negative). Must be given along with 'P0', 'rho' and 'g'.}

\item{P0}{is the air pressure in decibar or Pascal (see 'Pain').}

\item{rho}{is the mass density of the sea in kg pr m^3.}

\item{g}{is the gravitational acceleration in meters per s^2.}

\item{Pain}{is TRUE if pressure is given in Pascal and FALSE if given in decibar (10000 Pascal).}
}
\value{

}
\description{
Computes the speed of sound of the ocean for the given temperature, salinity and pressure values. Method adopted from "sw_svel.m" and based on UNESCO 1983.
}
\examples{
\dontrun{}

}
