% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_TSD.R
\name{subset_TSD}
\alias{subset_TSD}
\title{Extracts a subset of TSD data according to the array subset 'ind' and/or the cartesian subset 'range' and/or the logical/numeric vector of subscripts 'subset'.}
\usage{
subset_TSD(data, ind = list(), range = list(), subset = NULL,
  ind.out = FALSE, drop = TRUE, strict = TRUE, insert.NA = FALSE,
  only.match = FALSE, pad = c("end", "start"))
}
\arguments{
\item{data}{is a list of elements named according to the TSD file format.}

\item{ind}{is a list of indexes, as typed into the [] of an array, where 0 and NULL denotes all indexes.}

\item{range}{is a list of elements with names matching names if 'data', specifying the range of the corresponding elements.}

\item{subset}{is a numeric or logical vector/expression indicating elements or rows to keep (as used in subset()). Missing values are taken as false, and subset=0 or subset=NULL indicates no subsetting.}

\item{ind.out}{is TRUE if the indexes for the elements segmented are to be returned.}

\item{drop}{is TRUE if dimensions of only one level is to be removed from the output.}

\item{strict}{is TRUE if strict inequality is to be used when subsetting according to 'range'.}

\item{insert.NA}{is TRUE if the discarded data are to be kept as NA.}

\item{only.match}{is TRUE if only the arrays of length equal to the length of 'subset' are to be subsetted using 'subset'.}
}
\value{

}
\description{
Extracts a subset of TSD data according to the array subset 'ind' and/or the cartesian subset 'range' and/or the logical/numeric vector of subscripts 'subset'.
}
\examples{
\dontrun{}

}
