% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_TSD.R
\name{summary_TSD}
\alias{summary_TSD}
\title{Calculates variaous measures of mean sv, and total backscatter and volume.}
\usage{
summary_TSD(data, plot.hist = FALSE, minlen = 1, ...)
}
\arguments{
\item{data}{A list containing volume and the acoustic data eihter given as 'vbss' for a segment of the data or as 'vbsc'.}

\item{plot.hist}{Logical: If TRUE plot the histogram when calculating the robust mean sv.}

\item{minlen}{The minimum number of acoustic samples.}

\item{...}{Passed on to density() when calculating the robust mean sv.}
}
\description{
Calculates variaous measures of mean sv, and total backscatter and volume.
}
