% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySeg.R
\name{summarySeg}
\alias{summarySeg}
\title{(Internal) Returns a summary of segmentation data.}
\usage{
summarySeg(
  event,
  t,
  TS = -55,
  w = 0.369,
  segpar = 1:2,
  ang = NULL,
  maxRange = 500,
  n90 = 2
)
}
\value{

}
\description{
(Internal) Returns a summary of segmentation data.
}
\examples{
\dontrun{}

}
