% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv2Mask.R
\name{sv2MaskOnePing}
\alias{sv2MaskOnePing}
\title{Bayesian segmentation method for one ping}
\usage{
sv2MaskOnePing(sv, betaN, beta0, beta1, h, c, beams)
}
\arguments{
\item{sv}{A matrix of sv (linear values) for one ping.}

\item{betaN}{A vector of background noise for all beams.}

\item{beta0}{The lower schooling threshold given in linear value. Suggested value: 1e-5.}

\item{beta1}{The upper schooling threshold given in linear value. Suggested value: 1e-2.}

\item{h}{The standard deviation in the Gaussian kernel used to smooth the probability of "not-school". Suggested value: 5.}

\item{c}{The segmentation threshold. Suggested value: 1e-5.}

\item{beams}{A list of beam configuration data. Must contain he following elements: lenb, numb, plsl, asps, sint, absr, rofs. Read this from aa .beams file in the TSD format.}
}
\description{
Function to get the segmentation mask of one ping from the Bayesian segmentation method (Holmin et al., 2024)
}
