% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.circle.R
\name{track.circle}
\alias{track.circle}
\title{Returns a list of x-values and y-values of points defined by the fractions 'w' along the partial circular path defined by the list or matrix of origin points 'origin', the vector of radii 'r' and the list or matrix of angles 'ang'. Cumulative lengths of the path are also returned.}
\usage{
track.circle(origin, r, ang, w = NULL, dw = NULL, speed = NULL,
  normalized = FALSE, list.out = TRUE, plot = FALSE, ...)
}
\arguments{
\item{origin}{is the origin of the path, given as a list of two elements or a matrix of two columns, representing the x and y values of the origins.}

\item{r}{is a vector of length 'n' holding the input radius-variable.}

\item{ang}{is the input angle-variable as defined on the unit circle, given as a list of two elements or a matrix of two columns, representing the start and the end angles of the circle segments.}

\item{w}{are the positions along the path (x,y), or along the x-axis if alongx==TRUE. If 'w' is a single integer, it is set to seq(0,1,length.out=w). If 'speed' is given 'w' is interpreted as time values.}

\item{dw}{are the piecewise lengths along the path (x,y), or along the x-axis if alongx==TRUE, used only if w!=NULL. If 'dw' ihas length>1 it overrides 'w' by w=cumsum(dw).}

\item{speed}{is a vector of sound speeds of each circle segment. If given, 'w' is interpreted as time values starting from dt=0 at the start of the path.}

\item{normalized}{is TRUE if the positions given by 'w' (and 'speed') are or should be normalized to the set [0,1], where 1 represents the end of the path.}

\item{list.out}{is TRUE if the output should be a data frame with the names "x", "y" and "l". Else a matrix with the same names is returned.}

\item{plot}{is TRUE if a simple plot is to be drawn.}

\item{...}{are parameters to be passed on to plot() if plot==TRUE.}
}
\value{

}
\description{
Returns a list of x-values and y-values of points defined by the fractions 'w' along the partial circular path defined by the list or matrix of origin points 'origin', the vector of radii 'r' and the list or matrix of angles 'ang'. Cumulative lengths of the path are also returned.
}
\examples{
\dontrun{}

}
