% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl2rt.TSD.R
\name{vl2rt.TSD}
\alias{vl2rt.TSD}
\title{Transforms from velocity data to rotation data, given as counter clockwise around z-rotation and counter clockwise around x-rotation.}
\usage{
vl2rt.TSD(data, var = c("rtz", "rtx"), fresh = FALSE)
}
\arguments{
\item{data}{is a list containing the variables 'vlx*', 'vly*' and 'vlz*'.}

\item{var}{is a vector of the elements to return.}
}
\description{
Transforms from velocity data to rotation data, given as counter clockwise around z-rotation and counter clockwise around x-rotation.
}
\examples{
\dontrun{}

}
