% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volx1D_oneping.R
\name{volx1D_oneping}
\alias{volx1D_oneping}
\title{Calculates the volumes of the voxels of the EK60 multifrequency echosounder given the CTD-data and beams-dataof the particular data set 'dataset'. See read.event() for importing volumes of voxels.}
\usage{
volx1D_oneping(data)
}
\arguments{
\item{data}{is the list of the inputs used in the function as returned from read.TSD() (must include "").}
}
\value{

}
\description{
Calculates the volumes of the voxels of the EK60 multifrequency echosounder given the CTD-data and beams-dataof the particular data set 'dataset'. See read.event() for importing volumes of voxels.
}
\examples{
\dontrun{}

}
