% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volx2D_oneping.R
\name{volx2D_oneping}
\alias{volx2D_oneping}
\title{Calculates the volumes of the voxels of an omnidirectional fan of a multibeam sonar. Only the rectangular beams are treated!!! See read.event() for importing volumes of voxels.}
\usage{
volx2D_oneping(data, fanWidth = "b2")
}
\arguments{
\item{data}{is the list of the inputs used in the function as returned from read.TSD() (must include "").}

\item{fanWidth}{has a number of possible values (abbreviations allowed, so that the values "e" and "b" can be used):}
}
\value{

}
\description{
Calculates the volumes of the voxels of an omnidirectional fan of a multibeam sonar. Only the rectangular beams are treated!!! See read.event() for importing volumes of voxels.
}
\examples{
\dontrun{}

}
