% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volx3D_oneping.R
\name{volx3D_oneping}
\alias{volx3D_oneping}
\title{Calculates the volumes of the voxels of the MS70 multibeam sonar given the CTD-data and beams-data. See read.event() for importing volumes of voxels.}
\usage{
volx3D_oneping(data)
}
\arguments{
\item{data}{is the list of the inputs used in the function as returned from read.TSD() (must include "").}
}
\value{

}
\description{
Calculates the volumes of the voxels of the MS70 multibeam sonar given the CTD-data and beams-data. See read.event() for importing volumes of voxels.
}
\examples{
\dontrun{}

}
