% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel.edge.TSD.R
\name{voxel.edge.TSD}
\alias{voxel.edge.TSD}
\title{Calculates a matrix of edge points of voxels of an underwater acoustic system. The matrix contains r0 and r1, the closer and father range to each voxel; theta0 and theta1, the azimuth angle of the left and right dge of each voxel; and phi0 and phi1, the elevation angle of the lower and upper edge of each voxel, as seen from the vessel for a multi-beam sonar such as the Simrad MS70. !!!Only one time step is treated!!!!}
\usage{
voxel.edge.TSD(
  data,
  t = 1,
  esnm = NULL,
  seabed = -12000,
  rot = 1,
  compensation = c("pitch", "roll"),
  ideal = TRUE,
  stretch = 1,
  fanWidth = "b2",
  ...
)
}
\arguments{
\item{data}{is the list of inputs variables as returned from read.TSD (including ).}

\item{t}{is a single integer giving the time step for which the edge matrix in the coordinate system of the vessel is returned.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{seabed}{is the z-coordinate of the sea bed, usually provided by echo sounder data. Soundbeams reflected from the sea bed or the surface are reflected at the incidence angle.}

\item{rot}{is 1 if simple rotation using cosine and sine is to be used, and 2 if the function rotate() is to be used in the rotation. Times for the different methods (tested on MacBook Pro dual 2.8 GHz, 2010-02-09, with other applications running):}

\item{compensation}{is a vector of string giving which rotation values that are compensated for in the sonar. Only c("pitch","roll") is available for the current version. Used in soundbeam.TSD.}

\item{ideal}{is TRUE to represent the simple case where the speed of sound 'data$asps' is invariant of depth.}

\item{stretch}{is used to stretch the voxels of the ME70 multibeam echosounder in the direction of motion, so that space in between voxels is smoothed out.}

\item{fanWidth}{has a number of possible values: (1) "b1": one way beam width. (2) "b2": two way beam width. (3) "fe": beams modeled by rectangular cones with width withing the fan given by the inter-beam angle, and calculated using the equivalent beam angle. This normally causes larged fan width due to overlap between beams.}
}
\description{
Calculates a matrix of edge points of voxels of an underwater acoustic system. The matrix contains r0 and r1, the closer and father range to each voxel; theta0 and theta1, the azimuth angle of the left and right dge of each voxel; and phi0 and phi1, the elevation angle of the lower and upper edge of each voxel, as seen from the vessel for a multi-beam sonar such as the Simrad MS70. !!!Only one time step is treated!!!!
}
\examples{
\dontrun{}

}
