% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsc2netcdf4.R
\name{writeNetCDF4}
\alias{writeNetCDF4}
\alias{tsd2NetCDF4}
\alias{pings2NetCDF4}
\alias{seg2NetCDF4}
\title{Write data to a netCDF4 file}
\usage{
writeNetCDF4(data, con, units = NULL, dims = NULL)

tsd2NetCDF4(tsdFile, netCDF4dir, units = NULL, dims = NULL)

pings2NetCDF4(event)

seg2NetCDF4(event)
}
\arguments{
\item{data}{The list of data to write to the netCDF4 file.}

\item{con}{The path to the netCDF4 file.}

\item{units}{An optional list of units, named by the names of \code{data}.}

\item{dims}{An optional list of dimensions, named by the names of \code{data}, where each element must be a list of the elements Name and Unit as vectors of the names and units of the dimensions of the element of \code{data} that the dimensions correspond to (example: dims = list(vbsc = list(Name = c("sample", "beam", "ping"), Unit = c("SampleIndex", "BeamIndex", "PingIndex")))).}

\item{tsdFile}{The path to a TSD file.}

\item{netCDF4dir}{The path to the directory in which to put a netCDF4 file.}

\item{event}{The path to the event holding TSD files.}
}
\description{
\code{writeNetCDF4} Writes a list of data to a netCDF4 file.
\code{tsd2NetCDF4} Writes a list of data to a netCDF4 file.
\code{pings2NetCDF4} Convenience function for converting .pings files to netCDF4 files.
\code{seg2NetCDF4} Convenience function for converting .seg files to netCDF4 files.
}
