% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.getSchoolfileType.R
\name{echoIBM.getSchoolfileType}
\alias{echoIBM.getSchoolfileType}
\title{Function for extracting the type (dynamic or static) of the school files given by 'schoolfiles'. Four vectors are returned:
NA}
\usage{
echoIBM.getSchoolfileType(schoolfiles, dynschoolnames, staticschoolnames)
}
\arguments{
\item{schoolfiles}{is a vector of paths to school files.}

\item{dynschoolnames}{is a vector of four character strings representing dynamic school valiable names.}

\item{staticschoolnames}{is a vector of four character strings representing static school valiable names.}
}
\value{

}
\description{
Function for extracting the type (dynamic or static) of the school files given by 'schoolfiles'. Four vectors are returned:
NA
}
\examples{
\dontrun{}

}
