% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.specs.esnm.R
\name{get.specs.esnm}
\alias{get.specs.esnm}
\title{Returns specifications of the acoustic instrument 'esnm' regarding types of beams. The function treats single time steps, except for the output "region".}
\usage{
get.specs.esnm(data, var = c("rect", "offset", "region", "valid", "ind",
  "direction"), esnm = NULL, ind = NULL, add = 30, beamstypes = 1)
}
\arguments{
\item{data}{is a list of elements named according to the TSD file format holding the following information: "esnm" (may be given separately), "lenb", "", "", "", "", "", "", "", "", "", "", "".}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}
}
\value{

}
\description{
Returns specifications of the acoustic instrument 'esnm' regarding types of beams. The function treats single time steps, except for the output "region".
}
\examples{
\dontrun{}

}
