% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.files_caseInsensitive.R
\name{list.files_caseInsensitive}
\alias{list.files_caseInsensitive}
\title{Sorts in a case insensitive matter. Created to solve the problem of errors when listing files in R as run from the terminal, in which case sorting is not case insensitive. In the R-GUI sorting is done case insensitive.}
\usage{
list.files_caseInsensitive(path, pattern = NULL, all.files = FALSE,
  full.names = FALSE, recursive = FALSE, ignore.case = FALSE,
  method = "shell")
}
\arguments{
\item{x}{is an R-object as input to sort() / order().}

\item{decreasing}{used in sort() / order().}

\item{na.last}{used in sort() / order().}
}
\value{

}
\description{
Sorts in a case insensitive matter. Created to solve the problem of errors when listing files in R as run from the terminal, in which case sorting is not case insensitive. In the R-GUI sorting is done case insensitive.
}
\examples{
\dontrun{}

}
