% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.noise_file.R
\name{locate.noise_file}
\alias{locate.noise_file}
\title{(Internal) Locate noise file.}
\usage{
locate.noise_file(
  event,
  noisevar = c("bgns", "nrnp"),
  corvar = c("crb1", "olpb"),
  esnm = "MS70"
)
}
\arguments{
\item{event}{The path to the event.}

\item{noisevar}{A vector of noise variabes to find.}

\item{corvar}{A vector of noise correlation variabes to find.}

\item{esnm}{The name of the acoustic instrument for which to find the noise files.}
}
\value{

}
\description{
(Internal) Locate noise file.
}
\examples{
\dontrun{}

}
