% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.event_read_sgsc.R
\name{read.event_read_sgsc}
\alias{read.event_read_sgsc}
\title{Reads segmentation indices 'sgsc'.}
\usage{
read.event_read_sgsc(filelist, var, filesind, segpar, TIME, tlist, merge,
  msg = TRUE)
}
\arguments{
\item{filelist}{is a vector of all file names.}

\item{filesind}{is a vector of the indexes of the segmentation files in the list of files.}

\item{segpar}{is a list of elements named "bwGp", "lsth"/"rlst", "usth"/"rust", or "sgth"/"sgt0" specifying the parameters of the segmentation data to read.}

\item{TIME}{is the list returned from UNIX_time().}

\item{tlist}{is the list of time indexes to be read for each file.}
}
\value{

}
\description{
Reads segmentation indices 'sgsc'.
}
\examples{
\dontrun{}

}
